Fixed bugs, one more time... But the last one now !

+++ ft_strchr.c
@@ -6,7 +6,7 @@
 /*   By: dbiguene <dbiguene@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/09 09:54:09 by dbiguene          #+#    #+#             */
-/*   Updated: 2022/11/17 16:24:52 by dbiguene         ###   ########lyon.fr   */
+/*   Updated: 2022/11/18 19:18:21 by dbiguene         ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,7 @@ char	*ft_strchr(const char *s, int c)
 	i = -1;
 	while (s[++i] && s[i] != c % 256)
 		;
-	if (i == (size_t)ft_strlen(s) && c % 256 != 0)
+	if (i == ft_strlen(s) && c % 256 != 0)
 		return (NULL);
 	return ((char *)s + i);
 }
